feat: add ResponseEntity to all endpoints

+++ src/main/java/dev/leozinho/mmorpg/character/CharacterController.java
@@ -1,5 +1,7 @@
 package dev.leozinho.mmorpg.character;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,27 +18,47 @@ public class CharacterController {
 
     // Adicionar personagem (CREATE)
     @PostMapping("/create")
-    public CharacterDTO createCharacter(@RequestBody CharacterDTO character){
-        return service.createCharacter(character);
+    public ResponseEntity<?> createCharacter(@RequestBody CharacterDTO character){
+        CharacterDTO characterDTO = service.createCharacter(character);
+        return ResponseEntity.status(HttpStatus.CREATED).body(characterDTO);
     }
     // Listar todos os personagens(READ)
     @GetMapping("/all")
-    public List<CharacterDTO> showAllCharacters(){
-      return service.showAllCharacters();
+    public ResponseEntity<?> showAllCharacters(){
+        List<CharacterDTO> characters = service.showAllCharacters();
+        return  ResponseEntity.ok(characters);
     }
     // Listar personagem por id (READ)
     @GetMapping("/showCharacterById/{id}")
-    public CharacterDTO showCharacterById(@PathVariable Long id){
-      return service.showCharacterById(id);
+    public ResponseEntity<?> showCharacterById(@PathVariable Long id){
+        CharacterDTO character = service.showCharacterById(id);
+        if (character != null){
+           return ResponseEntity.ok(character);
+        }else {
+          return   ResponseEntity.status(HttpStatus.NOT_FOUND).body("Character o ID: " + id + " nao foi encontrado");
+        }
     }
     // Atualizar personagem (UPDATE)
     @PutMapping("/updateCharacterById/{id}")
-    public CharacterDTO updateCharacterById(@PathVariable Long id,@RequestBody CharacterDTO character){
-        return service.updateCharacterById(id, character);
+    public ResponseEntity<?> updateCharacterById(@PathVariable Long id,@RequestBody CharacterDTO character){
+        CharacterDTO characterDTO = service.showCharacterById(id);
+        if (characterDTO != null){
+          CharacterDTO characteUpdated = service.updateCharacterById(id, character);
+         return    ResponseEntity.ok(characteUpdated);
+        }else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Erro ao econtrar o chacater com o ID: " + id);
+        }
+
     }
     // Deletar personagem (DELETE)
     @DeleteMapping("/deleteCharacterById/{id}")
-    public void deleteCharacterById(@PathVariable Long id){
-        service.deleteCharacterById(id);
+    public ResponseEntity<?> deleteCharacterById(@PathVariable Long id){
+       CharacterDTO character = service.showCharacterById(id);
+       if (character != null){
+           service.deleteCharacterById(id);
+         return   ResponseEntity.ok("Character com o ID: " + id + " foi deletado com sucesso");
+       }else {
+          return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Falha ao procurar o Character com o ID: " + id);
+       }
     }
 }
